django_project/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
audioread==2.1.9
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
datasets==2.1.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
Django==4.0.4
djangorestframework==3.13.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.6.0
frozenlist==1.3.0
fsspec==2022.3.0
huggingface-hub==0.5.1
idna==3.3
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
librosa==0.9.1
llvmlite==0.38.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
multiprocess==0.70.12.2
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numba==0.55.1
numpy==1.21.6
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pooch==1.6.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.1.0
regex==2022.4.24
requests==2.27.1
resampy==0.2.2
responses==0.18.0
sacremoses==0.0.53
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
sentencepiece==0.1.96
six==1.16.0
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
terminado==0.13.3
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
torch==1.11.0
torchaudio==0.11.0
tornado==6.1
tqdm==4.64.0
traitlets==5.2.0
transformers==4.18.0
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
xxhash==3.0.0
yarl==1.7.2
